pyproject.toml
[tool.poetry]
name = "call-assistant"
version = "0.1.0"
description = ""
authors = ["Reinder Vos de Wael <reinder.vosdewael@childmind.org>"]
readme = "README.md"
packages = [{include = "call_assistant"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.1.2"
pydantic = "^2.4.2"
pygithub = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
